fixed x64 compilation problem
git-svn-id: https://only.mawhrin.net/repos/stuff/trunk@2076 62a0de91-87f2-0310-aeaa-8eaa5733e38e

+++ fbreader/src/formats/chm/LZXDecompressor.h
@@ -62,8 +62,8 @@ private:
 
 	DecodingState myState;
 
-	unsigned int myBlockSize;
-	unsigned int myBlockBytesLeft;
+	size_t myBlockSize;
+	size_t myBlockBytesLeft;
 
 	const unsigned int mySlotNumber;
 	HuffmanDecoder myMainTree;
